jcjones review

+++ aggregator_api/src/tests.rs
@@ -1548,7 +1548,7 @@ async fn post_taskprov_peer_aggregator() {
         aggregation_mode: Some(AggregationMode::Synchronous),
         collector_hpke_config: leader.collector_hpke_config().clone(),
         verify_key_init: *leader.verify_key_init(),
-        report_expiry_age: leader.report_expiry_age().map(|d| d.num_seconds()),
+        report_expiry_age: leader.report_expiry_age().map(|d| d.num_seconds() as u64),
         aggregator_auth_tokens: Vec::from(leader.aggregator_auth_tokens()),
         collector_auth_tokens: Vec::from(leader.collector_auth_tokens()),
     };
